Modification du calcul de la rampe de vitesse distance à tester. Plus modification de l'application du coef roue dans l'odométrie. (calcul sur des int innutiles jusqu'à présent)

+++ Carte Strat V2.X/codeurs.c
@@ -36,7 +36,7 @@ void ConfigQEI (void)
 	// QEI1CON
 	QEI1CONbits.CNTERR	= 0;				// Reset flag erreur sur compteur
 	QEI1CONbits.QEISIDL	= 0;				// Mode continu en idle
-    QEI1CONbits.QEIM        = 0b111;                        // Mode QEI avec X4
+    QEI1CONbits.QEIM    = 0b111;            // Mode QEI avec X4
 	QEI1CONbits.SWPAB	= 0;				// PA et PB non entrelacés
 	QEI1CONbits.PCDOUT	= 1;				// Position counter dir. status output active (IO controle par QEI)
 
@@ -59,7 +59,7 @@ void ConfigQEI (void)
 	// QEI2CON
 	QEI2CONbits.CNTERR	= 0;				// Reset flag erreur sur compteur
 	QEI2CONbits.QEISIDL	= 0;				// Mode continu en idle
-    QEI2CONbits.QEIM        = 0b111;                        // Mode QEI avec X4
+    QEI2CONbits.QEIM    = 0b111;            // Mode QEI avec X4
 	QEI2CONbits.SWPAB	= 0;				// PA et PB non intervertis
 	QEI2CONbits.PCDOUT	= 1;				// Position counter dir. status output active (IO controle par QEI)
 
@@ -112,7 +112,7 @@ void get_valeur_codeur (int codeur)
         res+= (int32_t) etat_overflow * max_codeur;
     }
 
-    POSITION[codeur].ecart = res * RESOLUTION_LOGICIELLE * (int32_t) sens_rotation;
+    POSITION[codeur].ecart = (double) (res * RESOLUTION_LOGICIELLE * (int32_t) (sens_rotation) );
     POSITION[codeur].ancien = POSITION[codeur].nouvelle;
 }
